fix: use type uintptr_t to avoid compilation issue

+++ src/show_alloc_mem.c
@@ -6,7 +6,7 @@
 /*   By: sganon <sganon@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/09/17 19:21:36 by sganon            #+#    #+#             */
-/*   Updated: 2018/09/17 21:36:38 by sganon           ###   ########.fr       */
+/*   Updated: 2018/09/17 21:49:29 by sganon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ void display_allocs(t_allocs *allocs, char *name)
 {
   char *base_addr;
 
-  base_addr = ft_itoa_base((unsigned int)allocs, 16);
+  base_addr = ft_itoa_base((uintptr_t)allocs, 16);
   to_upper_str(base_addr);
   ft_putstr(name);
   ft_putstr(" : 0x");
